tutorials/bulk-images/requirements.txt
bulk==0.1.0
keras==2.10.0
jupyterlab==3.4.8
embetter[vision]==0.2.0 
scikit-learn==1.1.2
umap-learn==0.5.3
